+++ src/hooks/useCartList.ts
@@ -1,0 +1,19 @@
+import { CartItem } from 'types/domain';
+import useUpdateCartItem from './useUpdateCartItem';
+
+export const useCartList = (cartList: CartItem[]) => {
+  const { toggleCartItemAllChecked, removeSelectedCartItem } = useUpdateCartItem(cartList);
+  const isAllItemChecked = cartList.every(cartItem => cartItem.checked);
+
+  const toggleCheckedAll = () => {
+    const targetItemList = cartList.filter(cartItem => cartItem.checked === isAllItemChecked);
+
+    toggleCartItemAllChecked(targetItemList);
+  };
+
+  const deleteSelectedItem = () => {
+    removeSelectedCartItem();
+  };
+
+  return { isAllItemChecked, toggleCheckedAll, deleteSelectedItem };
+};
